Conflicts:
	framework/Autoloader/package.xml
	framework/Autoloader/test/Horde/Autoloader/AllTests.php
	framework/Autoloader/test/Horde/Autoloader/phpunit.xml
	framework/Autoloader_Cache/lib/Horde/Autoloader/Cache.php
	framework/Autoloader_Cache/package.xml

+++ framework/Perms/lib/Horde/Perms/Permission/Kolab.php
@@ -1,8 +1,9 @@
 <?php
 /**
- * Maps folder permissions into the Horde_Permission system.
+ * Copyright 2006-2013 Horde LLC (http://www.horde.org/)
  *
- * PHP version 5
+ * See the enclosed file COPYING for license information (LGPL). If you
+ * did not receive this file, see http://www.horde.org/licenses/lgpl21.
  *
  * @category Horde
  * @package  Perms
@@ -14,21 +15,17 @@
 /**
  * Maps folder permissions into the Horde_Permission system.
  *
- * Copyright 2006-2012 Horde LLC (http://www.horde.org/)
- *
- * See the enclosed file COPYING for license information (LGPL). If you
- * did not receive this file, see http://www.horde.org/licenses/lgpl21.
- *
  * @category Horde
  * @package  Perms
  * @author   Gunnar Wrobel <wrobel@pardus.de>
  * @license  http://www.horde.org/licenses/lgpl21 LGPL 2.1
  * @link     http://pear.horde.org/index.php?package=Perms
  */
-class Horde_Perms_Permission_Kolab
-extends Horde_Perms_Permission
+class Horde_Perms_Permission_Kolab extends Horde_Perms_Permission
 {
-    /** Kolab ACL speak for all permissions on a shared object. */
+    /**
+     * Kolab ACL speak for all permissions on a shared object.
+     */
     const ALL = 'lrid';
 
     /**
@@ -36,14 +33,14 @@ extends Horde_Perms_Permission
      *
      * @var Horde_Perms_Permission_Kolab_Storage
      */
-    private $_storage;
+    protected $_storage;
 
     /**
      * The group handler.
      *
      * @var Horde_Group_Base
      */
-    private $_groups;
+    protected $_groups;
 
     /**
      * A cache for the folder acl settings. The cache holds the permissions
@@ -72,22 +69,45 @@ extends Horde_Perms_Permission
     }
 
     /**
-     * Gets the current permission of the folder and stores the values in the
-     * cache.
+     * Returns the permissions of the owner on this object.
      *
-     * @return NULL
+     * @since Horde_Perms 2.1.0
+     *
+     * @return integer  All owner permissions for this object.
+     */
+    public function getOwnerPermissions()
+    {
+        $users = parent::getUserPermissions();
+        return $users[$this->_storage->getOwner()];
+    }
+
+    /**
+     * Returns an array of all user permissions on this object.
+     *
+     * @param integer $perm  List only users with this permission level.
+     *                       Defaults to all users.
+     *
+     * @return array  All user permissions for this object, indexed by user.
+     */
+    public function getUserPermissions($perm = null)
+    {
+        $users = parent::getUserPermissions($perm);
+        unset($users[$this->_storage->getOwner()]);
+        return $users;
+    }
+
+    /**
+     * Returns the current permission of the folder and stores the values in
+     * the cache.
      */
     public function getCurrentPermissions()
     {
         $data = array();
-        /**
-         * @todo: Can we lazy load $this->data so that we restrict to using
+        /* @todo: Can we lazy load $this->data so that we restrict to using
          * MYRIGHTS only when that is all we need and use the full GETACL just
-         * when required.
-         */
+         * when required. */
         $acl = new Horde_Perms_Permission_Kolab_AclIterator(
-            $this->_storage->getAcl(),
-            $this->_storage->getOwner()
+            $this->_storage->getAcl()
         );
         foreach ($acl as $element) {
             $element->toHorde($data);
@@ -98,35 +118,37 @@ extends Horde_Perms_Permission
 
     /**
      * Saves the current permission values from the cache to the IMAP folder.
-     *
-     * @return NULL
      */
     public function save()
     {
-        /**
-         * @todo: If somebody else accessed the folder before us, we will
-         * overwrite the change here.
-         */
+        /* @todo: If somebody else accessed the folder before us, we will
+         * overwrite the change here. */
         $current = $this->getCurrentPermissions();
+        $owner = $this->_storage->getOwner();
 
         $elements = new Horde_Perms_Permission_Kolab_ElementIterator(
-            $this->data, $this->_groups, $this->_storage->getOwner()
+            $this->data, $this->_groups
         );
         foreach ($elements as $element) {
+            if ($owner == $element->getId()) {
+                continue;
+            }
             $this->_storage->setAcl($element->getId(), $element->fromHorde());
             $element->unsetInCurrent($current);
         }
 
         // Delete ACLs that have been removed
         $elements = new Horde_Perms_Permission_Kolab_ElementIterator(
-            $current, $this->_groups, $this->_storage->getOwner()
+            $current, $this->_groups
         );
         foreach ($elements as $element) {
+            if ($owner == $element->getId()) {
+                continue;
+            }
             $this->_storage->deleteAcl($element->getId());
         }
 
         // Load the permission from the folder again
         $this->data = $this->getCurrentPermissions();
     }
-
 }
